requirements.txt
Django==1.6.2
django-pipeline==1.3.20
-e git+https://github.com/vbabiy/django-pipeline-compass.git@f5a942380d0dfc3fafb8d867830780f4e576e7f9#egg=django_pipeline_compass
pyScss==1.2.0.post3
django-filebrowser==3.5.3
stripe==1.12.0


cssmin==0.2.0
slimit==0.8.1
django-grappelli==2.5.2
feedparser==5.1.3
django-localflavor==1.0

pillow==2.3.0

#DB
psycopg2==2.5.2
south==.0.8.4

#FORMS

#GENERAL TOOLS
-e hg+https://hg@bitbucket.org/kanja/django-registration@69e70e8fd55004fa18c8f234dad472daf5b55771#egg=django_registration
    #for styling admin changelist rows
-e git+git://github.com/litchfield/django-liststyle.git@master#egg=django-liststyle

django-annoying==0.7.9
django-extensions==1.3.3
django-model-utils==2.0
unipath==1.0
pytz==2013.9
manifesto==0.5

#REDIS CACHING
hiredis==0.1.2
redis==2.9.1
django-redis-cache==0.10.2
django-redis-sessions==0.4.0

## AWS
boto==2.24.0
django-storages==1.1.8
django-ses==0.4.1
#django_smtp_ssl

### TESTING
factory-boy==2.3.1
model_mommy==1.2

#HEROKU
dj-database-url==0.2.2
#dj-static==0.0.5
-e git+https://github.com/gone/dj-static.git@5da00529c033afa8e65aef13555b9eaec6dfc388#egg=dj-static
django-stripe-payments==2.0b34
django-toolbelt==0.0.1
ipython==1.1.0
gunicorn==18.0


django-jsonfield==0.9.12
django-mandrill==0.1.0
mandrill==1.0.53
jsmin

#cool stuff
django-reversion==1.8.0
python-social-auth==0.1.23


#OPTIONAL TOOLS
#whoosh
#django-haystack
